backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
pydantic==2.9.2
pydantic-settings==2.6.0
google-generativeai==0.8.3
sqlalchemy==2.0.36
python-dotenv==1.0.1
python-multipart==0.0.12
aiosqlite==0.20.0
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
